Fix alter table reorganize.
For offload tablespace map, postgresql
creates "temp" relation for reorganized
data. For yezzey, we need to know how
tablespace maps to external bucket, which is OTM entry
used for. Make sure that "old" and "new" relation refers
the same tablespace/buckets by simple in-memory hack.
This is actaully worked before, but was fooled at some point,
so add regression test for ALTER+OTM.